refactor: 화면 기준 패키지 재구성
