+++ app/src/main/java/nextstep/github/ui/screen/MainActivity.kt
@@ -1,9 +1,10 @@
-package nextstep.github
+package nextstep.github.ui.screen
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import nextstep.github.ui.screen.GithubApp
+import nextstep.github.MainApplication
+import nextstep.github.ui.screen.github.GithubApp
 import nextstep.github.ui.theme.GithubTheme
 
 class MainActivity : ComponentActivity() {
