+++ app/src/main/java/nextstep/github/ui/screen/github/GithubApp.kt
@@ -1,10 +1,9 @@
-package nextstep.github.ui.screen
+package nextstep.github.ui.screen.github
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
 import nextstep.github.di.AppContainer
-import nextstep.github.ui.viewmodel.GithubViewModel
 
 @Composable
 fun GithubApp(
@@ -14,4 +13,17 @@ fun GithubApp(
     val viewModel: GithubViewModel = viewModel(
         factory = GithubViewModel.provideFactory(appContainer.githubRepository)
     )
+
+    GithubScreen(
+        viewModel = viewModel,
+        modifier = modifier,
+    )
+}
+
+@Composable
+fun GithubScreen(
+    viewModel: GithubViewModel,
+    modifier: Modifier = Modifier
+) {
+
 }
